[✨feat] 메뉴바를 PC와 Mobile로 구분(VOTEGUARD-8)
- PC와 Mobile(사이즈)에 따라 보여질 메뉴바를 구분하기 위해 react-responsive를 설치
- PC와 Mobile에서 보여질 메뉴바를 각각의 컴포넌트로 나눠서 구현
- useResize 훅(커스텀훅)을 생성하여 반응형을 위한 기능 구현
- 아직 PC와 Mobile 둘 다 드롭다운을 완벽하게 꾸며내지 못 한 상태
