+++ Front/vote-guardian/package.json
@@ -13,6 +13,7 @@
     "node-sass": "^9.0.0",
     "react": "^18",
     "react-dom": "^18",
+    "react-responsive": "^9.0.2",
     "sass": "^1.70.0"
   },
   "devDependencies": {
