+++ Front/vote-guardian/src/components/VoteManage/VoteManageMenu/PcDropDown.scss
@@ -1,29 +1,22 @@
-html, body {
-    width: 100%;
-    height: 100%;
-    margin: 0;
+//투표
+.drop-wrap {
+    margin: 3vmin 0;
+    .drop-title {
+        width: 100%;
+        height: 6vmin;
+        p {
+            margin: 0;
+            font-size: 1.6vw;
+            line-height: 6vmin;
+        }
+        &.true, .drop-title:hover {
+            background-color: rgba( #4A83EF, 0.5 );
+            font-weight: bold;
+        }
+    }
 }
 
-@media (min-width:700px) and (max-width:1920px) {
-    //메뉴바 컴포넌트의 가장 바깥 div
-    .bar-wrap {
-        width: 12em;
-        height: 100%;
-        border-right: 3px solid #4A83EF;
-        box-sizing: border-box;
-        text-align: center;
-        position: relative;
-
-        //투표 가디언 타이틀
-        .header {
-            //글자
-            font-size: 30px;
-            line-height: 2em;
-            color: white;
-            //박스
-            height: 2em;
-            background-color: #4A83EF;
-        }
+/*
         //드롭다운 메뉴
         .drop-menu {
             margin-top: 2em;
@@ -48,7 +41,7 @@ html, body {
         .logo-wrap {
             width: 100%;
             height: 9em;
-            position: absolute;
+            // position: absolute;
             bottom: 0;
             img {
                 width: 100%;
@@ -56,5 +49,5 @@ html, body {
             }
         }
     }
-}
 
+*/
